* IupCoreChangesCMake:
  CMake: New files from Iup resync
  SymbolVisibility: iup_varg.h
  Official: revision 4693
  CMake: Had problems finding webkitgtk on Raspbian Jessie. This contains additional fallbacks to look for other packages.
  SymbolVisibility: Needed for va_list functions.
  CMake: Trying older version of these CMake files to require lower version. Taken from 2.8
  SymbolVisibility: iupluaweb needs to export symbols, but the place to put it is in the header. But the use case never imports the header. Furthermore, the code generator only seems to allow one header to be specified. So the change is to put the export markers in the header, and change it to include iupluaweb.h instead of iupweb.h. Then make iupluaweb.h include iupweb.h
  CMake: Build iuplua, iupluaweb as library instead of module
  IupCoreChanges: More ifdefs needed to avoid getting GTK as default for new platforms.
  SymbolVisibility: Symbols needed for iuplua iupweblua
  CMake: Adding IupLua support + iupweblua
  Revert "IupCoreChanges: iupweb: introduces GOBACK, GOFORWARD to GTK & Windows. CANGOFORWARD, CANGOBACK for GTK."
  IupCoreChanges: iupweb: introduces GOBACK, GOFORWARD to GTK & Windows. CANGOFORWARD, CANGOBACK for GTK.
  CMake: I screwed up a commit or merge somewhere. This adds back the Windows OLE/Web stuff.
  CMake: new Android file for iupweb
  CMake: Changes to support building iupweb on Android (plus tweaks for Apple). Also passing -Wno-invalid-source-encoding
  CMake: Adding Windows iupole

+++ srcscintilla/iupsci_tab.c
@@ -43,7 +43,7 @@ SCI_GETHIGHLIGHTGUIDE
 static char* iScintillaGetTabSizeAttrib(Ihandle *ih)
 {
   int widthInChars;
-  widthInChars = IupScintillaSendMessage(ih, SCI_GETTABWIDTH, 0, 0);
+  widthInChars = (int)IupScintillaSendMessage(ih, SCI_GETTABWIDTH, 0, 0);
   return iupStrReturnInt(widthInChars);
 }
 
@@ -62,7 +62,7 @@ static int iScintillaSetTabSizeAttrib(Ihandle *ih, const char *value)
 
 static char* iScintillaGetHighlightGuideAttrib(Ihandle *ih)
 {
-  int col = IupScintillaSendMessage(ih, SCI_GETHIGHLIGHTGUIDE, 0, 0);
+  int col = (int)IupScintillaSendMessage(ih, SCI_GETHIGHLIGHTGUIDE, 0, 0);
   return iupStrReturnInt(col);
 }
 
@@ -79,7 +79,7 @@ static int iScintillaSetHighlightGuideAttrib(Ihandle *ih, const char *value)
 
 static char* iScintillaGetIndentationGuidesAttrib(Ihandle *ih)
 {
-  int indentView = IupScintillaSendMessage(ih, SCI_GETINDENTATIONGUIDES, 0, 0);
+  int indentView = (int)IupScintillaSendMessage(ih, SCI_GETINDENTATIONGUIDES, 0, 0);
   char* str[] = {"NONE", "REAL", "LOOKFORWARD", "LOOKBOTH"};
   return str[indentView];
 }
@@ -110,7 +110,7 @@ static int iScintillaSetUseTabsAttrib(Ihandle* ih, const char* value)
 
 static char* iScintillaGetUseTabsAttrib(Ihandle* ih)
 {
-  return iupStrReturnBoolean (IupScintillaSendMessage(ih, SCI_GETUSETABS, 0, 0)); 
+  return iupStrReturnBoolean((int)IupScintillaSendMessage(ih, SCI_GETUSETABS, 0, 0)); 
 }
 
 void iupScintillaRegisterTab(Iclass* ic)
